LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/report-to-both_coop-ro.https-expected.txt


FAIL Access from opener assert_equals: expected (string) "coop" but got (undefined) undefined
FAIL Access from openee assert_equals: expected (string) "coop" but got (undefined) undefined
